Mark changes in existing build script
